+++ src/cks/__init__.py
@@ -76,7 +76,7 @@ from .interface import (
 
 _EXTERNAL_COUNT = load_external_constraints()
 
-__version__ = "1.9.0"
+__version__ = "1.9.1"
 
 VERSION = tuple(
     int(part)